[UpDate]Bat
몬스터 피격 이펙트 추가.
-> 박쥐 : 공격 애니메이션 삭제

+++ Assets/Manager/IDamged.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamged
 {
-    public void TakeDamage(int damage); 
+    public void TakeDamage(float damage); 
 }
